compilation with/without 64
