+++ src/curve25519/fe/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod load;
+
 #[cfg(not(all(
     any(target_arch = "x86", target_arch = "x86_64"),
     target_feature = "sse4.1"
