+++ src/ed25519.rs
@@ -21,7 +21,6 @@ use crate::curve25519::{curve25519, ge_scalarmult_base, sc_muladd, sc_reduce, Fe
 use crate::digest::Digest;
 use crate::sha2::Sha512;
 use core::convert::TryFrom;
-use core::ops::{Add, Mul, Sub};
 
 #[deprecated(since = "0.4.0", note = "use `PRIVATE_KEY_LENGTH`")]
 pub const SEED_LENGTH: usize = 32;
@@ -248,11 +247,11 @@ pub fn exchange(public_key: &[u8; 32], private_key: &[u8; PRIVATE_KEY_LENGTH]) -
 
 fn edwards_to_montgomery_x(ed_y: &Fe) -> Fe {
     let ed_z = &Fe::ONE;
-    let temp_x = ed_z.add(ed_y);
-    let temp_z = ed_z.sub(ed_y);
+    let temp_x = ed_z + ed_y;
+    let temp_z = ed_z - ed_y;
     let temp_z_inv = temp_z.invert();
 
-    let mont_x = temp_x.mul(&temp_z_inv);
+    let mont_x = &temp_x * &temp_z_inv;
 
     mont_x
 }
